Updated ddiskdesc.asm and rb1773desc.asm so would work with lwtools.
Updated level1/modules/ddiskdesc.asm so that if default
values are not given by assembler define command statement
then definitions are set to a default value.

Updated level1/modules/rb1773desc.asm so that if default
values are not given by assembler define command statement
then definitions are set to a default value.