cohr: Support both COVDG control codes and COHR Esc codes
$0B is the only control character that is used by both protocols.
I decided to keep the the COVDG command as this protocol is more common.
So far I've been able to find only 3 OS9 applications that are using the
CoHR drivers, all by Dragon Data: Stylograph, Dynacalc and RMS. Only RMS
seems to use the $0B (Home) control code, but it is easy to patch the
RMS.trm file to replace $0B with the $01 CoVDG Home command. So this
seems like a reasonable sacrifice.