sys/makefiles: Remove broken prerequisite from makefile rule
The $@ automatic variable is not allowed in a rule prerequisite.
It could be added using secondary expansion, but here there is
no point anyway, since the makefile is not meant to build this
prerequisite - it should exist in any case.

Also, since the coco3_6309/sys/makefile file is identical to the one
in coco3/sys, just keep the latter and include it from the other.