sierra: Correct set.pri.base routine used in KingsQuest4
point X to priority table

The procedure updates the priority table, pointed by reg X but this
register is never assigned. Reg X point to the command table ($81B0) when
the procedure is executed and overwrites it, crashing the interpreter
almost immediately after the procedure returns. I found out that the
priority table resides at $05EE. This table is an array of 168 8-bit
values that defines a priority number between 1 and 15 for each line of
the graphic screen.

I had to insert another routine at the end of the module to have reg X
to point at the priority table location and modify the command table to
call the alternate routine.

Patch by Guillaume Major and Robert Gault.