mnln.asm: Fix up "move to (x,y)" after last commit
There was still a little bug when the motion type is "move to (x.y)". In
this case, the location of the formatted message is in reg U, but since
I replaced "push u" by "push x" at L0C71, the wrong value is pushed on
the stack. So the instruction "tfr u,x" needs to be inserted just before
reaching L0C71.

But since the code is position dependent, we can't do this. So we add
a small patch stub to the end of the module.

Analysis and original patch by Guillaume Major.
Rewritten patch by Tormod Volden.