sierra: Fix clear.text.rect command in mnln.asm
There was a problem with arguments pushed onto the stack before calling
a subroutine. The fix is 4 bytes shorter than the original code so an
fcb directive is added to keep the following code at the same address.


The syntax of the clear.text.rect command is:

clear.text.rect(byt row1, byt col1, byt row2, byt col2, byt color);

All parameters are 8-bit values. row1 and row2 can be between 0 and 23,
col1 and col2 can be between 0 and 39.